+++ src/app/page.tsx
@@ -1,10 +1,10 @@
+import { useEffect } from "react";
 import RootLayout from "./layout";
 import Introduction from "./component/intro/Introduction";
 import CommonBenefits from "./component/membershipBenefits/CommonBenefits";
 import MembershipBenefits from "./component/membershipBenefits/MembershipBenefits";
-import MembershipPrice from "./component/membershipPrice/MembershipPrice";
-import { useEffect } from "react";
-import SponsorsReview from "./component/sponsorsReview/SponsorsReview";
+import MembershipPrice from "./component/membershipPayment/membershipPrice/MembershipPrice";
+import SponsorsReview from "./component/membershipPayment/sponsorsReview/SponsorsReview";
 // import * as styles from "./MainPage.css";
 const MainPage = () => {
   useEffect(() => {
